alterações de layout e criação de um diretório para as funções
